+++ skymp5-client/src/index.ts
@@ -40,6 +40,7 @@ import { LoadOrderVerificationService } from "./services/services/loadOrderVerif
 import { BrowserService } from "./services/services/browserService";
 import { AuthService } from "./services/services/authService";
 import { NetInfoService } from "./services/services/netInfoService";
+import { AnimDebugService } from "./services/services/animDebugService";
 
 once("update", () => {
   Utility.setINIBool("bAlwaysActive:General", true);
@@ -83,7 +84,8 @@ const main = () => {
       new LoadOrderVerificationService(sp, controller),
       new BrowserService(sp, controller),
       new AuthService(sp, controller),
-      new NetInfoService(sp, controller)
+      new NetInfoService(sp, controller),
+      new AnimDebugService(sp, controller)
     ];
     SpApiInteractor.setup(listeners);
     listeners.forEach(listener => SpApiInteractor.registerListenerForLookup(listener.constructor.name, listener));
